+++ sem/parse_test.go
@@ -10,6 +10,45 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func Test_DefaultParser(t *testing.T) {
+	v, err := DefaultParser([]byte(`1.2.3`), 0)
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v)
+	assert.NoError(t, err)
+	v, err = DefaultParser([]byte(`v1.2.3`), 0)
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v)
+	assert.NoError(t, err)
+	v, err = DefaultParser([]byte(`x`), 0)
+	assert.Zero(t, v)
+	assert.Error(t, err)
+	v, err = DefaultParser([]byte(`1.2.3`), RuleDisableParseTag)
+	assert.Equal(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "",
+		Build:      "",
+	}, v)
+	assert.NoError(t, err)
+	v, err = DefaultParser([]byte(`v1.2.3`), RuleDisableParseTag)
+	assert.Zero(t, v)
+	assert.Error(t, err)
+	v, err = DefaultParser([]byte(`x`), RuleDisableParseTag)
+	assert.Zero(t, v)
+	assert.Error(t, err)
+}
+
 func Test_ParseVersion(t *testing.T) {
 	testParseVersion(t, ParseVersion)
 }
@@ -104,4 +143,45 @@ func testParseTag(t *testing.T, f func([]byte) (Ver, error)) {
 func Test_Parse(t *testing.T) {
 	testParseVersion(t, Parse)
 	testParseTag(t, Parse)
+}
+
+func assertUnmarshalText(t *testing.T, expected Ver, input string, f form) {
+	t.Helper()
+	v, err := unmarshalText("x", []byte(input), f)
+	assert.Equal(t, expected, v)
+	assert.NoError(t, err)
+}
+
+func assertUnmarshalTextFail(t *testing.T, error, input string, f form) {
+	t.Helper()
+	v, err := unmarshalText("x", []byte(input), f)
+	assert.Zero(t, v)
+	assert.EqualError(t, err, `sem.x: `+error)
+}
+
+func Test_unmarshalText(t *testing.T) {
+	MaxTextLength = 0
+	assertUnmarshalText(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "x",
+		Build:      "y",
+	}, `1.2.3-x+y`, formVersion)
+	assertUnmarshalText(t, Ver{
+		Major:      1,
+		Minor:      2,
+		Patch:      3,
+		PreRelease: "x",
+		Build:      "y",
+	}, `v1.2.3-x+y`, formTag)
+	assertUnmarshalTextFail(t, `invalid version`, ``, 0)
+	assertUnmarshalTextFail(t, `"1.2.3": expected tag form`, `1.2.3`, formTag)
+	assertUnmarshalTextFail(t, `"v1.2.3": tag form not allowed`, `v1.2.3`, formVersion)
+	assertUnmarshalTextFail(t, `"1000000000000000000000000000000.2.3": invalid major`, `1000000000000000000000000000000.2.3`, formVersion)
+	assertUnmarshalTextFail(t, `"1.2000000000000000000000000000000.3": invalid minor`, `1.2000000000000000000000000000000.3`, formVersion)
+	assertUnmarshalTextFail(t, `"1.2.3000000000000000000000000000000": invalid patch`, `1.2.3000000000000000000000000000000`, formVersion)
+	MaxTextLength = 4
+	assertUnmarshalTextFail(t, `input too long (5 > 4)`, `xxxxx`, 0)
+	MaxTextLength = 0
 }