+++ size/parse.go
@@ -22,7 +22,7 @@ const (
 )
 
 var (
-	// MaxTextLength allows limiting UnmarshalText input.
+	// MaxTextLength allows limiting Parser input.
 	// Set 0 to disable this setting.
 	MaxTextLength = 128
 
@@ -30,104 +30,126 @@ var (
 	// Set 0 to disable this setting.
 	MaxObjectKeys = 16
 
-	// DisableUnmarshalTextUnit allows disabling unit specification at DefaultUnmarshalText.
-	// If true, DefaultUnmarshalText returns error if Size is presented with specified unit.
-	DisableUnmarshalTextUnit = false
+	// Parser is used by Size.UnmarshalText and Size.UnmarshalJSON functions.
+	Parser = DefaultParser
 
-	// DisableUnmarshalJSONStringForm allows disabling string form at DefaultUnmarshalJSON.
-	// If true, DefaultUnmarshalJSON returns error if Size is presented as JSON string.
-	DisableUnmarshalJSONStringForm = false
-
-	// DisableUnmarshalJSONObjectForm allows disabling object form at DefaultUnmarshalJSON.
-	// If true, DefaultUnmarshalJSON returns error if Size is presented as JSON object.
-	DisableUnmarshalJSONObjectForm = false
-
-	// UnmarshalText is used by Size.UnmarshalText and DefaultUnmarshalJSON functions.
-	UnmarshalText = DefaultUnmarshalText
-
-	// UnmarshalJSON is used by Size.UnmarshalJSON function.
-	UnmarshalJSON = DefaultUnmarshalJSON
+	errUnitDisabled                 = errors.New("unit disabled")
+	errExpectedNumberStringOrObject = errors.New("expected number, string or object")
+	errObjectFormDisabled           = errors.New("object form disabled")
+	errStringFormDisabled           = errors.New("string form disabled")
+	errMissingValueKey              = errors.New("missing value key")
+	errMissingUnitKey               = errors.New("missing unit key")
+	errDuplicatedValueKey           = errors.New("duplicated value key")
+	errDuplicatedUnitKey            = errors.New("duplicated unit key")
 )
 
-// DefaultUnmarshalText parse Size from text.
-// Allowed forms are number or string (with or without specified unit).
-// Spaces before and after data are allowed and ignored.
+type (
+	// Rule allows configuring Parser behavior.
+	Rule int
+)
+
+const (
+	// RuleDisableUnit allows disabling unit specification.
+	// If present, Parser returns error if Size is presented with specified unit.
+	RuleDisableUnit = Rule(1 << iota)
+
+	// RuleEnableJSONStringForm allows JSON string form.
+	// If not present, Parser returns error if Size is presented as JSON string.
+	RuleEnableJSONStringForm
+
+	// RuleEnableJSONObjectForm allows JSON object form.
+	// If not present, Parser returns error if Size is presented as JSON object.
+	RuleEnableJSONObjectForm
+
+	ruleIsJSON            = RuleEnableJSONStringForm | RuleEnableJSONObjectForm
+	ruleUnmarshalTextMask = RuleDisableUnit
+
+	defaultParserFuncName = "DefaultParser"
+)
+
+var (
+	// DefaultRule is used by Size.UnmarshalText and Size.UnmarshalJSON converting functions.
+	DefaultRule = RuleEnableJSONStringForm | RuleEnableJSONObjectForm
+)
+
+// DefaultParser parse Size from input.
+// Allowed forms are number, string (with or without specified unit) or JSON object (depends on passed rule value).
+//
+// For number and string form, spaces before and after data are allowed and ignored.
 // Also, spaces/non-breakable-spaces/underscores are allowed and ignored between digits and also as number/unit separator.
 //
-// See also MaxTextLength and DisableUnmarshalTextUnit.
-func DefaultUnmarshalText(data []byte) (Size, error) {
-	const funcName = "DefaultUnmarshalText"
-
+// JSON object form must contain key "value" with positive or zero number value and "unit" with string value and valid size unit.
+// JSON object keys are case-insensitive.
+//
+// See also MaxTextLength and MaxObjectKeys.
+func DefaultParser(data []byte, r Rule) (Size, error) {
 	if l := len(data); MaxTextLength != 0 && l > MaxTextLength {
 		// do not use input for "input too long" error
-		return 0, newParseError(funcName, "", fmt.Errorf("input too long (%d > %d)", l, MaxTextLength))
+		return 0, newParseError(defaultParserFuncName, "", fmt.Errorf("input too long (%d > %d)", l, MaxTextLength))
 	}
 
+	if r&ruleIsJSON != 0 {
+		return unmarshalJSON(data, r)
+	}
+
+	return unmarshalText(data, r)
+}
+
+func unmarshalText(data []byte, r Rule) (Size, error) {
 	s := string(data)
 	number, unit := prepareNumber(s)
 
 	if number == "" {
-		return 0, newParseError(funcName, s, nil)
+		return 0, newParseError(defaultParserFuncName, s, nil)
 	}
 	value, err := strconv.ParseUint(number, 10, 64)
 	if err != nil {
-		return 0, newParseError(funcName, s, err)
+		return 0, newParseError(defaultParserFuncName, s, err)
 	}
 
 	if unit == "" {
 		return Size(value), nil
 	}
-	if DisableUnmarshalTextUnit {
-		return 0, newParseError(funcName, s, errors.New("unit disabled"))
+	if r&RuleDisableUnit != 0 {
+		return 0, newParseError(defaultParserFuncName, s, errUnitDisabled)
 	}
 
 	size, err := New(value, unit)
 	if err != nil {
-		return 0, newParseError(funcName, s, err)
+		return 0, newParseError(defaultParserFuncName, s, err)
 	}
 	return size, nil
 }
 
-// DefaultUnmarshalJSON parse Size from JSON.
-// Allowed forms are number, string (with or without specified unit) or object.
-//
-// Strings are parsed with DefaultUnmarshalText.
-//
-// Object must contain key "value" with positive or zero number value and "unit" with string value and valid size unit.
-// Object keys are case-insensitive.
-//
-// See also MaxObjectKeys, DisableUnmarshalJSONStringForm and DisableUnmarshalJSONObjectForm.
-func DefaultUnmarshalJSON(data []byte) (Size, error) {
-	const funcName = "DefaultUnmarshalJSON"
-
+func unmarshalJSON(data []byte, r Rule) (Size, error) {
 	d := json.NewDecoder(bytes.NewReader(data))
 	d.UseNumber()
 	t, err := d.Token()
 	if err != nil {
-		return 0, newParseError(funcName, string(data), err)
+		return 0, newParseError(defaultParserFuncName, string(data), err)
 	}
 	switch v := t.(type) {
 	case json.Delim:
 		if v != '{' {
-			return 0, newParseError(funcName, string(data), errors.New("expected number, string or object"))
+			return 0, newParseError(defaultParserFuncName, string(data), errExpectedNumberStringOrObject)
 		}
-		if DisableUnmarshalJSONObjectForm {
-			return 0, newParseError(funcName, string(data), errors.New("object form disabled"))
+		if r&RuleEnableJSONObjectForm == 0 {
+			return 0, newParseError(defaultParserFuncName, string(data), errObjectFormDisabled)
 		}
 		size, err := unmarshalJSONObject(d)
 		if err != nil {
-			return 0, newParseError(funcName, string(data), err)
+			return 0, newParseError(defaultParserFuncName, string(data), err)
 		}
 		return size, nil
 	case json.Number:
-		return UnmarshalText([]byte(v))
+		return unmarshalText([]byte(v), 0)
 	case string:
-		if DisableUnmarshalJSONStringForm {
-			return 0, newParseError(funcName, string(data), errors.New("string form disabled"))
+		if r&RuleEnableJSONStringForm == 0 {
+			return 0, newParseError(defaultParserFuncName, string(data), errStringFormDisabled)
 		}
-		return UnmarshalText([]byte(v))
+		return unmarshalText([]byte(v), 0)
 	default:
-		return 0, newParseError(funcName, string(data), fmt.Errorf("unexpected type %T", t))
+		return 0, newParseError(defaultParserFuncName, string(data), fmt.Errorf("unexpected type %T", t))
 	}
 }
 
@@ -181,7 +203,7 @@ keys:
 		switch strings.ToLower(key) {
 		case ObjectKeyValue:
 			if value != nil {
-				return 0, errors.New("duplicated value key")
+				return 0, errDuplicatedValueKey
 			}
 			value, err = decodeValue(d)
 			if err != nil {
@@ -192,7 +214,7 @@ keys:
 			}
 		case ObjectKeyUnit:
 			if unit != nil {
-				return 0, errors.New("duplicated unit key")
+				return 0, errDuplicatedUnitKey
 			}
 			unit, err = decodeUnit(d)
 			if err != nil {
@@ -212,10 +234,10 @@ keys:
 
 func newOrError(value *uint64, unit *string) (Size, error) {
 	if value == nil {
-		return 0, errors.New("missing value key")
+		return 0, errMissingValueKey
 	}
 	if unit == nil {
-		return 0, errors.New("missing unit key")
+		return 0, errMissingUnitKey
 	}
 	return New(*value, *unit)
 }